All R resources neede for NorScir at Rapporteket

An implemetation of all R related resources for NorScir at Rapporteket, including R-functions for analysis and plotting, noweb-files and even artwork needed for producing R-dependent reports

Sample data

The package contains a set of sample data that can be used for development purposes. Not to be used in a PRODUCTION environment!

Functions

Standard R-functions for processing registry data

Noweb-files

LaTeX docs weaved with R-code to produce reports. These are accessed by system.file, E.g system.file("NSsamelDok.Rnw", package = "NorScir"). You may process the noweb files directly using Sweave or knitr, e.g knitr::knit(system.file("NSsamleDok.Rnw", package = "NorScir")), or by using the wrapper of the rapbase package, e.g. rapbase::RunNoweb(nowebFileName, packageName, weaveMethod = "knitr").

Artwork

Artwork is accessed in the same way as for the noweb-files, e.g system.file("NorSCIR_logo_blue_norsk.pdf", package = "NorScir")

Examples

data(NorScirSampleData)
names(NorScirSampleData)
#>   [1] "InjuryDt"          "AdmitDt"           "DischgDt"         
#>   [4] "ANeuExmDt"         "FNeuExmDt"         "QolDt"            
#>   [7] "AdmitRehDt"        "CNeuExmDt"         "InjuryDateUnknown"
#>  [10] "PasientId"         "isMale"            "AlderAar"         
#>  [13] "ReshId"            "ShNavn"            "DataSett"         
#>  [16] "RelatedID"         "SkjemaID"          "DatasettGuid"     
#>  [19] "RapportgrunnlagID" "KontaktID"         "DataSettID"       
#>  [22] "Scietiol"          "isVrtbrInj"        "isAssocInj"       
#>  [25] "isSpnlSurg"        "VentAssi"          "PlaceDis"         
#>  [28] "ASensLvlAreaL"     "ASensLvlLC"        "ASensLvlLT"       
#>  [31] "ASensLvlLL"        "ASensLvlLS"        "ASensLvlAreaR"    
#>  [34] "ASensLvlRC"        "ASensLvlRT"        "ASensLvlRL"       
#>  [37] "ASensLvlRS"        "AMtrLvlAreaL"      "AMtrLvlLC"        
#>  [40] "AMtrLvlLT"         "AMtrLvlLL"         "AMtrLvlLS"        
#>  [43] "AMtrLvlAreaR"      "AMtrLvlRC"         "AMtrLvlRT"        
#>  [46] "AMtrLvlRL"         "AMtrLvlRS"         "AAis"             
#>  [49] "FSensLvlAreaL"     "FSensLvlLC"        "FSensLvlLT"       
#>  [52] "FSensLvlLL"        "FSensLvlLS"        "FSensLvlAreaR"    
#>  [55] "FSensLvlRC"        "FSensLvlRT"        "FSensLvlRL"       
#>  [58] "FSensLvlRS"        "FMtrLvlAreaL"      "FMtrLvlLC"        
#>  [61] "FMtrLvlLT"         "FMtrLvlLL"         "FMtrLvlLS"        
#>  [64] "FMtrLvlAreaR"      "FMtrLvlRC"         "FMtrLvlRT"        
#>  [67] "FMtrLvlRL"         "FMtrLvlRS"         "FAis"             
#>  [70] "Skjematype"        "CSensLvlAreaL"     "CSensLvlLC"       
#>  [73] "CSensLvlLT"        "CSensLvlLL"        "CSensLvlLS"       
#>  [76] "CSensLvlAreaR"     "CSensLvlRC"        "CSensLvlRT"       
#>  [79] "CSensLvlRL"        "CSensLvlRS"        "CMtrLvlAreaL"     
#>  [82] "CMtrLvlLC"         "CMtrLvlLT"         "CMtrLvlLL"        
#>  [85] "CMtrLvlLS"         "CMtrLvlAreaR"      "CMtrLvlRC"        
#>  [88] "CMtrLvlRT"         "CMtrLvlRL"         "CMtrLvlRS"        
#>  [91] "CAis"              "ANeuNoMeasure"     "FNeuNoMeasure"    
#>  [94] "CNeuNoMeasure"     "OutOfHosptlDy"     "HosptlDy"         
#>  [97] "SatGenrl"          "SatPhys"           "SatPsych"         
#> [100] "DagerRehab"        "DagerTilRehab"     "NevrNivaaInn"     
#> [103] "NevrNivaaUt"       "NevrNivaaKtr"     
if (FALSE) { # \dontrun{
NSFigAndeler(RegData = NorScirSampleData, valgtVar = "Alder", reshID = 123456, hentData=0)
knitr::knit(system.file("NSsamleDok.Rnw", package = "NorScir"))
} # }